Updated tests for grunt dev tasks.
